refactor(workspace): split actions.rs god-file into actions/ by family (CODE-4)
Mechanically split the 5154-LOC workspace/actions.rs into an actions/ directory grouped by handler family (connections, settings, audit, metrics, documents, scripts, query, charts_dashboards). Pure code movement: signatures and bodies unchanged; the cross-family pub(super) methods became pub(in crate::ui::views::workspace) to keep the exact original visibility now that they live one module deeper. Updated the governance acceptance test to read the actions/ module directory instead of the single file. No behavior change; all workspace tests pass.

All 3951 workspace tests pass; clippy -D warnings and rustfmt clean.
