+++ crates/dbflux/tests/mcp_ui_governance_acceptance_tests.rs
@@ -6,6 +6,29 @@ fn read_workspace_file(relative_path: &str) -> String {
     fs::read_to_string(root.join(relative_path)).expect("file should be readable")
 }
 
+/// Reads every `.rs` file under a module directory and concatenates them.
+///
+/// Used for source-wiring assertions against modules that are split into a
+/// directory of sibling files, so the checks stay agnostic to which file a
+/// given symbol lives in.
+fn read_workspace_module(relative_dir: &str) -> String {
+    let root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+    let dir = root.join(relative_dir);
+
+    let mut entries: Vec<PathBuf> = fs::read_dir(&dir)
+        .expect("module directory should be readable")
+        .map(|entry| entry.expect("dir entry should be readable").path())
+        .filter(|path| path.extension().is_some_and(|ext| ext == "rs"))
+        .collect();
+    entries.sort();
+
+    entries
+        .iter()
+        .map(|path| fs::read_to_string(path).expect("file should be readable"))
+        .collect::<Vec<_>>()
+        .join("\n")
+}
+
 #[test]
 fn ui_contains_trusted_client_and_connection_policy_controls() {
     let mcp_settings = read_workspace_file("../dbflux_ui_windows/src/settings/mcp_section.rs");
@@ -26,7 +49,7 @@ fn ui_contains_trusted_client_and_connection_policy_controls() {
 #[test]
 fn ui_contains_approval_and_audit_controls_with_workspace_wiring() {
     let governance_view = read_workspace_file("../dbflux_ui_document/src/governance.rs");
-    let workspace_actions = read_workspace_file("../dbflux_ui/src/ui/views/workspace/actions.rs");
+    let workspace_actions = read_workspace_module("../dbflux_ui/src/ui/views/workspace/actions");
     let workspace_dispatch = read_workspace_file("../dbflux_ui/src/ui/views/workspace/dispatch.rs");
     let workspace_mod = read_workspace_file("../dbflux_ui/src/ui/views/workspace/mod.rs");
 
@@ -48,7 +71,7 @@ fn ui_contains_approval_and_audit_controls_with_workspace_wiring() {
 #[test]
 fn audit_workspace_actions_retarget_existing_document_and_close_governance_overlay() {
     let audit_document = read_workspace_file("../dbflux_ui_document/src/audit/mod.rs");
-    let workspace_actions = read_workspace_file("../dbflux_ui/src/ui/views/workspace/actions.rs");
+    let workspace_actions = read_workspace_module("../dbflux_ui/src/ui/views/workspace/actions");
 
     assert!(audit_document.contains("pub fn set_category_filter"));
     assert!(audit_document.contains("doc.pending_initial_load = false;"));
